% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classify.R
\name{classify.aggregate}
\alias{classify.aggregate}
\title{Primitive Bagging Sentiment Classification}
\usage{
classify.aggregate(sentences, pos.words, neg.words, .progress = "none")
}
\arguments{
\item{sentences}{a vector of sentences to classify}

\item{pos.words}{the dictionary of positive words}

\item{neg.words}{the dictionary of negative words}

\item{.progress}{boolean value indicating whether a progress bar should be displayed}
}
\description{
Provides a simple aggregation function in order to clasify a vector of sentences.
Incorporates text trimming and removes extraneous characters for better analysis.
}
\examples{
classify.aggregate(c("I am happy"), c("happy", "overjoyed"), c("sad", "depressed"), .progress='none')
}

