% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classify.R
\name{classify.dtm}
\alias{classify.dtm}
\title{Creates a Document/Term Matrix}
\usage{
classify.dtm(sentences, language = "english", minDocFreq = 1,
  minWordLength = 4, removeNumbers = TRUE, removePunctuation = TRUE,
  removeStopwords = TRUE, stemWords = FALSE, stripWhitespace = TRUE,
  toLower = TRUE, weighting = weightTf)
}
\arguments{
\item{sentences}{a character vector of sentences to use for training}

\item{language}{the language to use for word stemming}

\item{minDocFreq}{the minimum number of times a word is needed in a document to be included
in the analysis}

\item{minWordLength}{the minimnum word length for inclusion in analysis}

\item{removeNumbers}{a boolean regarding whether numbers whould be removed}

\item{removeStopwords}{a boolean specifying whether stopwords in the specified language should be removed}

\item{stemWords}{a boolean specying whether words should be stemmed to their root form in the language specified}

\item{stripWhitespace}{a boolean indicating whether whitespace should be stripped}

\item{toLower}{a boolean indicating whether all words should be transformed to their lowercase representations}

\item{weighting}{either weightTf or weightIfIdf (see the package \pkg{tm} for details)}

\item{removePuncutation}{a boolean specifying whether to remove puncutation}
}
\description{
Provides a utility function used to create a document/term matrix using
the package \pkg{tm}
}
\examples{
classify.dtm(c("I am happy", "I am sad", "I am miserable", "The weather is good today"))
}

