% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classify.R
\name{classify.viralheat}
\alias{classify.viralheat}
\title{ViralHeat API Sentiment Classification}
\usage{
classify.viralheat(text, api.key)
}
\arguments{
\item{text}{a vector of sentences to classify}

\item{api.key}{your API key for authentication}
}
\description{
Exposes ViralHeat's API in order to calculate the sentiment of a vector of text
based on their predefined algorithms. Text is passed as a JSON parameter to the
website, so internet connection is required for successful completion. Note that
the maximum number of characters is 360 (due to ViralHeat constraints). An API
key is required.
}

