% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridSearch.R
\name{gridSearch2D}
\alias{gridSearch2D}
\title{Grid search for errors}
\usage{
gridSearch2D(var_dflt_h, var_dflt_l, sent, type = "all",
  mean.sent = mean(sent), sd.sent = sd(sent), min.observ = 60,
  grid.length = 50, prob.lim = c(0, 1))
}
\value{
value1

A matrix of the infile
}
\description{
Some description - I do not have time for that
}
