% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.multiplot}
\alias{plot.multiplot}
\title{Multiplot (ggplot) Utility}
\usage{
\method{plot}{multiplot}(..., plotlist = NULL, cols)
}
\arguments{
\item{plotlist}{the list of plots}

\item{\dots}{the plots to stitch together}

\item{the}{columns for the plot}
}
\description{
Provides a utility to plot multiple ggplot instances together
}

