% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process.R
\name{process.stemwords}
\alias{process.stemwords}
\title{Word Stemming Utility}
\usage{
process.stemwords(document)
}
\arguments{
\item{document}{a vector of sentences to stem}
}
\description{
Provides a word stemming utility function for use in the naive
bayes classifier.
}

