% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process.R
\name{process.striplinks}
\alias{process.striplinks}
\title{Stip Text Links Utility}
\usage{
process.striplinks(x)
}
\arguments{
\item{x}{the text to process}
}
\description{
Strips 'http*' links from text in order for cleaner analysis
}
\examples{
process.striplinks("http://google.com is a search engine brought to you by Google")
}

