% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape.IMDb}
\alias{scrape.IMDb}
\title{IMDb Scraper Utility}
\usage{
scrape.IMDb(url)
}
\arguments{
\item{url}{the url to scrape}
}
\description{
Provides a simple scraping utility for access to the IMDb movie database
}
\examples{
scrape.IMDb('http://www.imdb.com/title/tt0903747/epdate') # Scrapes Breaking Bad Ratings
}

