% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape.retrieveklout}
\alias{scrape.retrieveklout}
\title{Klout Analysis for Twitter Utility}
\usage{
scrape.retrieveklout(handles, api.key, na_omit = TRUE, .progress = "none")
}
\arguments{
\item{handles}{a character vector of twitter handles}

\item{api.key}{a twitter API key}

\item{na_omit}{a boolean regarding whether to omit 'NA' values}

\item{.progress}{boolean value indicating whether a progress bar should be displayed}
}
\description{
Interfaces with klout to determine the influential rating of any twitter user
}

