\name{sentR-package}
\alias{sentR-package}
\alias{sentR}
\docType{package}
\title{
Provides varied techniques for big data sentiment analysis in R
}
\description{
Provides a more holistic understanding of sentiment analysis with numerous implemented methodologies including Naive bayes, Score Aggregation, and external database sourcing.
}
\details{
\tabular{ll}{
Package: \tab sentR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-03-11\cr
License: \tab MIT \cr
}
Includes core sentiment processing functions:

aggregate
naivebayes
viralheat
And exposes processing and analyzing utilities:

stemwords
striplinks
multiplot
Provides scraping functionality (basic):

IMDb
retrieveklout
}
\author{
Manan Shah
}

\keyword{ package, sentiment, analysis }
